More changes to satisfy splint. Passes the -weak tests pretty well now, with a few warnings because of 32/64 bit differences.

+++ btree.c
@@ -15,7 +15,7 @@ extern long Ticks;
 //===================================================================
 struct BTreeNode * CreateBTreeNode(int key, void * data)
 {
-	struct BTreeNode * node = (struct BTreeNode *)AllocUMem((long)sizeof(struct BTreeNode));
+	struct BTreeNode * node = (struct BTreeNode *)AllocUMem(sizeof(struct BTreeNode));
 	node->key = key;
 	node->data = data;
 	node->lesser = 0;
